code cleanup

+++ libfreeipmi/src/api/ipmi-lan-interface-api.h
@@ -29,9 +29,9 @@ int ipmi_lan_cmd (ipmi_ctx_t ctx,
 
 int ipmi_lan_cmd_raw (ipmi_ctx_t ctx,
                       const void *buf_rq,
-                      size_t buf_rq_len,
+                      unsigned int buf_rq_len,
                       void *buf_rs,
-                      size_t buf_rs_len);
+                      unsigned int buf_rs_len);
 
 int ipmi_lan_2_0_cmd (ipmi_ctx_t ctx,
                       fiid_obj_t obj_cmd_rq,
@@ -39,9 +39,9 @@ int ipmi_lan_2_0_cmd (ipmi_ctx_t ctx,
 
 int ipmi_lan_2_0_cmd_raw (ipmi_ctx_t ctx,
                           const void *buf_rq,
-                          size_t buf_rq_len,
+                          unsigned int buf_rq_len,
                           void *buf_rs,
-                          size_t buf_rs_len);
+                          unsigned int buf_rs_len);
 
 #endif /* ipmi-lan-interface.h */
